Implement an RPCFlush() interface for the RPC client
This version of go-codec (compared to the old keybase fork) performs
a Flush() after encoding every object.  This means that, when sending
an RPC request, there are 3 flushes: one for the RPC Request object,
one for the RPC arguments, and another one right after that (but there
were no unflushed writes before it).

This had an unfortunate effect on our RPC client implementation, which
sent each Write() as a separate message with tag RPCRequestTag, and
thus this single request turned into 2 requests -- the first missing
the arguments, and the second one malformed altogether.

In this change, we introduce an RPCFlush() method for the connection
used by the RPC codec.  RPCFlush() is called when the entire RPC request
and arguments are written.  This will allow our RPC client to send them
together as appropriate (until we eventually fix it by moving to RPC
over HTTP).